requirements.txt
requests
pyyaml
tabulate
python-dotenv
